Refactored according to new file layout proposed
