+++ providers/domainnameshop/domainnameshopProvider.go
@@ -3,13 +3,25 @@ package domainnameshop
 import (
 	"encoding/json"
 	"fmt"
-	"strconv"
-	"strings"
 
 	"github.com/StackExchange/dnscontrol/v3/models"
-	"github.com/StackExchange/dnscontrol/v3/pkg/diff"
 	"github.com/StackExchange/dnscontrol/v3/providers"
 )
+
+/**
+
+DomainNameShop Provider
+
+Info required in 'creds.json':
+	- token		API Token
+	- secret	API Secret
+
+*/
+
+type domainNameShopProvider struct {
+	Token  string // The API token
+	Secret string // The API secret
+}
 
 var features = providers.DocumentationNotes{
 	providers.CanAutoDNSSEC:          providers.Cannot(),        // Maybe there is support for it
@@ -23,16 +35,10 @@ var features = providers.DocumentationNotes{
 	providers.CanUseSOA:              providers.Cannot(),        // Does not seem to support it
 	providers.CanUseSRV:              providers.Can(),
 	providers.CanUseSSHFP:            providers.Cannot(),        // Does not seem to support it
-	providers.CanUseTLSA:             providers.Unimplemented(), // //Seems to support but needs to be implemented
+	providers.CanUseTLSA:             providers.Unimplemented(), // Seems to support but needs to be implemented
 	providers.DocCreateDomains:       providers.Unimplemented(), // Not tested
 	providers.DocDualHost:            providers.Unimplemented(), // Not tested
 	providers.DocOfficiallySupported: providers.Cannot(),
-}
-
-// dnsimpleProvider is the handle for this provider.
-type domainNameShopProvider struct {
-	Token  string // The API token
-	Secret string // The API secret
 }
 
 // Register with the dnscontrol system.
@@ -44,10 +50,6 @@ func init() {
 	}
 
 	providers.RegisterDomainServiceProviderType("DOMAINNAMESHOP", fns, features)
-}
-
-func auditRecords(records []*models.RecordConfig) error {
-	return nil
 }
 
 // newDomainNameShopProvider creates a DomainNameShop specific DNS provider.
@@ -67,121 +69,30 @@ func newDomainNameShopProvider(conf map[string]string, metadata json.RawMessage)
 	return api, nil
 }
 
-func (api *domainNameShopProvider) GetNameservers(domain string) ([]*models.Nameserver, error) {
-	ns, err := api.getNS(domain)
-	if err != nil {
-		return nil, err
-	}
-	return models.ToNameservers(ns)
+func auditRecords(records []*models.RecordConfig) error {
+	return nil
 }
 
-func (api *domainNameShopProvider) GetDomainCorrections(dc *models.DomainConfig) ([]*models.Correction, error) {
-	dc.Punycode()
-	existingRecords, err := api.GetZoneRecords(dc.Name)
-	if err != nil {
-		return nil, err
-	}
-
-	// Normalize
-	models.PostProcessRecords(existingRecords)
-
-	// Merge TXT strings to one string
-	for _, rc := range dc.Records {
-		if rc.HasFormatIdenticalToTXT() {
-			rc.SetTargetTXT(strings.Join(rc.TxtStrings, ""))
-		}
-	}
-
-	// Domainnameshop doesn't allow arbitrary TTLs they must be a multiple of 60.
-	for _, record := range dc.Records {
-		record.TTL = fixTTL(record.TTL)
-	}
-
-	differ := diff.New(dc)
-	_, create, delete, modify, err := differ.IncrementalDiff(existingRecords)
-	if err != nil {
-		return nil, err
-	}
-
-	var corrections = []*models.Correction{}
-
-	// Delete record
-	for _, r := range delete {
-		domainID := r.Existing.Original.(*domainNameShopRecord).DomainID
-		recordID := strconv.Itoa(r.Existing.Original.(*domainNameShopRecord).ID)
-
-		corr := &models.Correction{
-			Msg: fmt.Sprintf("%s, record id: %s", r.String(), recordID),
-			F:   func() error { return api.deleteRecord(domainID, recordID) },
-		}
-		corrections = append(corrections, corr)
-	}
-
-	// Create records
-	for _, r := range create {
-		domainName := strings.Replace(r.Desired.GetLabelFQDN(), r.Desired.GetLabel()+".", "", -1)
-		dnsR, err := api.fromRecordConfig(domainName, r.Desired)
-		if err != nil {
-			return nil, err
-		}
-		corr := &models.Correction{
-			Msg: r.String(),
-			F:   func() error { return api.CreateRecord(domainName, dnsR) },
-		}
-
-		corrections = append(corrections, corr)
-	}
-
-	for _, r := range modify {
-		domainName := strings.Replace(r.Desired.GetLabelFQDN(), r.Desired.GetLabel()+".", "", -1)
-
-		dnsR, err := api.fromRecordConfig(domainName, r.Desired)
-
-		if err != nil {
-			return nil, err
-		}
-
-		dnsR.ID = r.Existing.Original.(*domainNameShopRecord).ID
-
-		corr := &models.Correction{
-			Msg: r.String(),
-			F:   func() error { return api.UpdateRecord(dnsR) },
-		}
-
-		corrections = append(corrections, corr)
-	}
-
-	return corrections, nil
+type domainResponse struct {
+	ID          int      `json:"id"`
+	Domain      string   `json:"domain"`
+	Nameservers []string `json:"nameservers"`
 }
 
-func (api *domainNameShopProvider) GetZoneRecords(domain string) (models.Records, error) {
-	records, err := api.getDNS(domain)
-	if err != nil {
-		return nil, err
-	}
-
-	var existingRecords []*models.RecordConfig
-	for i := range records {
-		rC := toRecordConfig(domain, &records[i])
-		existingRecords = append(existingRecords, rC)
-	}
-
-	return existingRecords, nil
-}
-
-const minAllowedTTL = 60
-const maxAllowedTTL = 604800
-const multiplierTTL = 60
-
-func fixTTL(ttl uint32) uint32 {
-	// if the TTL is larger than the largest allowed value, return the largest allowed value
-	if ttl > maxAllowedTTL {
-		return maxAllowedTTL
-	} else if ttl < 60 {
-		return minAllowedTTL
-	}
-
-	// Return closest rounded down possible
-
-	return (ttl / multiplierTTL) * multiplierTTL
+type domainNameShopRecord struct {
+	ID             int    `json:"id"`
+	Host           string `json:"host"`
+	TTL            uint16 `json:"ttl,omitempty"`
+	Type           string `json:"type"`
+	Data           string `json:"data"`
+	Priority       string `json:"priority,omitempty"`
+	ActualPriority uint16
+	Weight         string `json:"weight,omitempty"`
+	ActualWeight   uint16
+	Port           string `json:"port,omitempty"`
+	ActualPort     uint16
+	CAATag         string `json:"tag,omitempty"`
+	ActualCAAFlag  string `json:"flags,omitempty"`
+	CAAFlag        uint64
+	DomainID       string
 }
