--- providers/domainnameshop/structs.go
@@ -1,25 +1,0 @@
-package domainnameshop
-
-type domainResponse struct {
-	ID          int      `json:"id"`
-	Domain      string   `json:"domain"`
-	Nameservers []string `json:"nameservers"`
-}
-
-type domainNameShopRecord struct {
-	ID             int    `json:"id"`
-	Host           string `json:"host"`
-	TTL            uint16 `json:"ttl,omitempty"`
-	Type           string `json:"type"`
-	Data           string `json:"data"`
-	Priority       string `json:"priority,omitempty"`
-	ActualPriority uint16
-	Weight         string `json:"weight,omitempty"`
-	ActualWeight   uint16
-	Port           string `json:"port,omitempty"`
-	ActualPort     uint16
-	CAATag         string `json:"tag,omitempty"`
-	ActualCAAFlag  string `json:"flags,omitempty"`
-	CAAFlag        uint64
-	DomainID       string
-}
